requirements.txt
# langchain
# langchain-astradb
# langchain-huggingface
# langchain-groq
# langchain-community
# python-dotenv
# pandas
# fastapi
# uvicorn
# python-multipart

aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0   
anyio==4.10.0
astrapy==2.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0        
certifi==2025.8.3        
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7  
deprecation==2.1.0       
distro==1.9.0
dnspython==2.7.0
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.7.0        
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45        
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-astradb==0.6.1
langchain-community==0.3.29
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.10
langsmith==0.4.21
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.2.0
numpy==2.3.2
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
streamlit==1.49.1
tenacity==9.1.2
tokenizers==0.22.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
watchdog==6.0.0
yarl==1.20.1
zstandard==0.24.0